% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA_segments_maker.R
\name{AA_segments_maker}
\alias{AA_segments_maker}
\title{Individual Amino Acid Position Dataframe for HLA loci}
\usage{
AA_segments_maker(loci)
}
\arguments{
\item{loci}{HLA locus, or loci.}
}
\value{
A dataframe for a locus with individual amino acid position information for all alleles for that locus. If multiple loci are input for the parameter, a list will be returned with corresponding dataframes for each locus. Note -- ANHIG/IMGT-HLA alignment sequence data begins at a negative enumeration for exon 1, coding for the leader peptide, which is eventually cleaved off; positive enumeration is reserved for exons that encode alpha and beta domains. In addition, the alignment enumeration skips counting inDels. An actual sequence enumeration was devised to combat potential confusion from starting at a negative enumeration, and enumerating positions with inDels. Column headers for amino acid position enumeration are based off the actual sequence -- the first row of the dataframe represents corresponding amino acid positions for the alignment sequence positions.
}
\description{
Extracts alignment sequence information for a given locus from the ANHIG/IMGTHLA Github Repository to produce a dataframe, which contains individual amino acid data for each amino acid position for all alleles, for a user-defined HLA locus or loci. The first 4 columns are locus, allele, trimmed allele, and allele_name.
}
\examples{
#AA_segments_maker with one locus as input
AA_segments_maker("C")

#AA_segments_maker with multiple loci as input
AA_segments_maker(c("A", "B", "C"))
}
