% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_HLAalignments}
\alias{all_HLAalignments}
\title{Version 3.38.0 Protein Alignments for all HLA loci}
\format{A list containing HLA alignments in a dataframe format for each locus.}
\source{
https://github.com/ANHIG/IMGTHLA/tree/Latest/alignments
}
\usage{
data(all_HLAalignments)
}
\description{
A list object containing protein alignments for all HLA loci. Alignments were downloaded from the ANHIG/IPD-IMGT Github respository.
}
\keyword{datasets}
