% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLAASD.R
\name{BLAASD}
\alias{BLAASD}
\title{BLAASD - Build Loci Amino Acid Specific Dataframe}
\usage{
BLAASD(loci)
}
\arguments{
\item{loci}{HLA locus, or loci.}
}
\value{
A dataframe for a locus with individual amino acid position information for all alleles for that locus. If multiple loci are input for the parameter, a list will be returned with corresponding dataframes for each locus.
}
\description{
Extracts alignment sequence information for a given locus from the ANHIG/IMGTHLA Github Repository to produce a dataframe, which contains individual amino acid data for each amino acid position for all alleles, for a user-defined HLA locus or loci. The first 4 columns are locus, allele, trimmed allele, and allele_name.
}
\examples{
#BLAASD with one locus as input
BLAASD("C")

#BLAASD with multiple loci as input
BLAASD(c("A", "B", "C"))
}
