% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCoordinates.R
\name{convertCoordinates}
\alias{convertCoordinates}
\title{Converts longitude and latitude with cardinal directions to proper coordinates}
\usage{
convertCoordinates(heatmapdata)
}
\arguments{
\item{heatmapdata}{Manipulated Solberg dataset produced from dataSubset.}
}
\value{
Solberg dataset with appropriate latitude and longitude enumerations
}
\description{
The Solberg dataset contains some coordinates associated with a cardinal direction, such as 50S. This function converts those coordinates to their enumerations -- in this case, 50S is -50. North (above equator) and East (east of prime meridian) positions are positive, and South (below equator) and West (west of prime meridian) are negative.
}
