% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countSpaces.R
\name{countSpaces}
\alias{countSpaces}
\title{Counts whitespaces in a string}
\usage{
countSpaces(x)
}
\arguments{
\item{x}{A string.}
}
\value{
Number of white spaces between elements of a split string.
}
\description{
Provided string is split, and white spaces are counted in between split elements. Written by Josh Bredeweg (reddit user jbraids1421).
}
\note{
This function was written found on Reddit and written by Josh Bredeweg, user jbraids1421.
}
\examples{
countSpaces("The quick  brown   fox jumped    over  the lazy   dog.")
}
