% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMotif.R
\name{findMotif}
\alias{findMotif}
\title{Narrows AA_segments table to only alleles that have the amino acid motif of interest}
\usage{
findMotif(motif)
}
\arguments{
\item{motif}{An amino acid motif in the following format: Locus*##$~##$~##$, where ## identifies a peptide position, and $ identifies an amino acid residue. Motifs can include any number of amino acids.}
}
\value{
A dataframe containing a subset of the amino-acid alignment produced from BLAASD(), with only alleles that contain the user-defined motif. If the motif is not find, a dataframe is returned, where one column has the motif, and the other column contains an error message.
}
\description{
Isolates the individual amino acid position dataframe produced by AA_segments_maker to only alleles with the user-defined motif. If the user-defined motif does not correspond to any alleles, an error message is output.
}
\examples{

#example with actual motif
findMotif("DRB1*26F~28E~30Y")
findMotif("DRB1*26F~28E")

#example with non-existent motif
findMotif("DRB1*26F~28E~30Z")

#extracting names of alleles with user-defined motif
findMotif("DRB1*26F~28E~30Y")[,4]
}
