% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_finder.R
\name{motif_finder}
\alias{motif_finder}
\title{Narrows AA_segments table to only alleles that have the amino acid motif of interest}
\usage{
motif_finder(motif)
}
\arguments{
\item{motif}{an amino acid motif in the following format: Locus*AA_1*AA_2*AA_3}
}
\value{
a dataframe with alleles that have the user input amino acid motif
}
\description{
Isolates the individual amino acid position dataframe produced by AA_segments_maker to only alleles with the user-defined motif. If the user-defined motif does not correspond to any alleles, an error message is output.
}
\examples{

#example with actual motif
motif_finder("DRB1*26F~28E~30Y")
motif_finder("DRB1*26F~28E")

#example with non-existent motif
motif_finder("DRB1*26F~28E~30Z")
}
