% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PALM.R
\name{PALM}
\alias{PALM}
\title{Population Allele Locating Mapmaker}
\usage{
PALM(gdataset, motif, color = TRUE, filterMigrant = TRUE)
}
\arguments{
\item{gdataset}{The Solberg dataset, a comma-separated value (CSV) file, which is the '1-locus-alleles.dat' file in the results.zip archive at http://pypop.org/popdata/. See the vignette for more details on pertinent parts of the Solberg dataset used for this package.}

\item{motif}{An amino acid motif in the following format: Locus*##$~##$~##$, where ## identifies a peptide position, and $ identifies an amino acid residue. Motifs can include any number of amino acids.}

\item{color}{A logical parameter that identifies if the heat maps should be made in color (TRUE) or gray scale (FALSE). The default option is TRUE.}

\item{filterMigrant}{A logical parameter that determines if admixed populations (OTH) and migrant populations (i.e. any complexities with the 'mig') should be excluded from the dataset. The default option is TRUE.}
}
\value{
None, but the produced heatmap is written to the user's working directory as a .jpg file, where the default filename is 'basemap.jpg'.
}
\description{
Produces a heatmap from the manipulated Solberg dataset by using the Generic Mapping Tools (GMT) R Package, which is an interface between R and the GMT Map-Making software. GMT commands are called via a Bash script through using the gmt.system() function in the GMT R package.
}
