% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countSpaces.R
\name{countSpaces}
\alias{countSpaces}
\title{Counts whitespaces in a string}
\usage{
countSpaces(x)
}
\arguments{
\item{x}{A string.}
}
\value{
A numeric vector that identifies the number of spaces between each set of non-space characters in the input string.
}
\description{
Returns the number of spaces between each set of non-space characters in a string.
}
\note{
This function was obtained from Reddit and written by Josh Bredeweg, user jbraids1421.
}
\examples{
countSpaces("The quick  brown   fox jumped    over  the lazy   dog.")
}
