% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSubset.R
\name{dataSubset}
\alias{dataSubset}
\title{Solberg dataset manipulation}
\usage{
dataSubset(dataset, motif)
}
\arguments{
\item{dataset}{The Solberg dataset, a comma-separated value (CSV) file, which is the '1-locus-alleles.dat' file in the results.zip archive at http://pypop.org/popdata/. See the vignette for more details on pertinent parts of the Solberg dataset used for this package.}

\item{motif}{An amino acid motif in the following format: Locus*##$~##$~##$, where ## identifies a peptide position, and $ identifies an amino acid residue. Motifs can include any number of amino acids.}
}
\value{
A dataframe format of the Solberg dataset, where the rows are ordered by population name, and limited to populations that have information on the locus of interest.
}
\description{
Manipulates the Solberg dataset by adding in a new column containing locus*allele information, reordering the dataset based on population name, and subsetting the data to only the locus of interest.
}
\note{
For internal SSHAARP use only.
}
