% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSubset.R
\name{dataSubset}
\alias{dataSubset}
\title{Solberg dataset manipulation}
\usage{
dataSubset(dataset, motif)
}
\arguments{
\item{dataset}{Solberg dataset. Can be found at http://pypop.org/popdata/ under results.zip.}

\item{motif}{An amino acid motif in the following format: Locus*AA_1*AA_2*AA_3}
}
\value{
The Solberg dataset ordered by population name, and limited to populations that have information on the locus of interest.
}
\description{
Manipulates the Solberg dataset by adding in a new column containing locus*allele information, reordering the dataset based on population name, and subsetting the data to only the locus of interest.
}
