% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PALM.R
\name{PALM}
\alias{PALM}
\title{Population Allele Locating Mapmaker}
\usage{
PALM(gdataset, motif, color = TRUE, filterMigrant = TRUE)
}
\arguments{
\item{gdataset}{The Solberg dataset, a comma-separated value (CSV) file, which is the '1-locus-alleles.dat' file in the results.zip archive at http://pypop.org/popdata/. Columns of the Solberg dataset that are pertinent to this package, along with an explanation of their contents, are as follows: popname - contains the population name, followed by a year, indicating the year of publication of the literature the population information was derived from; contin - contains three letter abbreviations of 10 major world regions. See the vignette for abbreviations. Significantly admixed populations were placed in an OTH group; complex - assigns level of complexity, ranging from 1 (least complex) to 3 (most complex), which estimates the degree of potential admixture in a population sample. Complexities with 'mig' indicate migrant populations; latit - latitude of the continent the population is found in; longit - longitude of the continent the population is found in; locus - HLA locus information of collected data; allele_v3 - HLA allele information of collected data; allele.freq - frequency of the corresponding HLA allele in a continent for a population.}

\item{motif}{An amino acid motif in the following format: Locus*##$##$##$, where ## identifies a peptide position, and $ identifies an amino acid residue. Motifs can include any number of amino acids.}

\item{color}{A logical parameter that identifies if the heat maps should be made in color (TRUE) or gray scale (FALSE). The default option is TRUE.}

\item{filterMigrant}{A logical parameter on whether OTH (admixed populations) and migrant populations should be filtered out of the dataset. The default option is TRUE.}
}
\value{
A .jpg file of the produced heatmap is output to the user's working directory.
}
\description{
Produces a heatmap from the manipulated Solberg dataset by using the Generic Mapping Tools (GMT) R Package, which is an interface between R and the GMT Map-Making software. GMT commands are called via a Bash script through using the gmt.system() function in the GMT R package.
}
