% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PALM.R
\name{PALM}
\alias{PALM}
\title{Population Allele Locating Mapmaker}
\usage{
PALM(motif, filename = SSHAARP::solberg_dataset, color = TRUE,
  filterMigrant = TRUE)
}
\arguments{
\item{motif}{An amino acid motif in the following format: Locus*##$~##$~##$, where ## identifies a peptide position, and $ identifies an amino acid residue. Motifs can include any number of amino acids.}

\item{filename}{The filename of the local copy of the Solberg dataset - the defaulted filename is the solberg_dataset in the SSHAARP package.}

\item{color}{A logical parameter that identifies if the heat maps should be made in color (TRUE) or gray scale (FALSE). The default option is TRUE.}

\item{filterMigrant}{A logical parameter that determines if admixed populations (OTH) and migrant populations (i.e. any complexities with the 'mig') should be excluded from the dataset. The default option is TRUE.}
}
\value{
None, but the produced heatmap is written to the user's working directory as a .jpg file, where the filename is "'the input motif'.jpg".
}
\description{
Produces a frequency heatmap for a specified amino-acid motif, based on the allele frequency data in the Solberg dataset.
}
\note{
The produced frequency heatmap is generated by using the Generic Mapping Tools (GMT) R Package, which is an interface between R and the GMT Map-Making software.

The Solberg dataset is the tab-delimited ‘1-locus-alleles.dat’ text file in the results.zip archive at http://pypop.org/popdata/.

The Solberg dataset is also prepackaged into SSHAARP as 'solberg_dataset'.
}
\examples{
#example to produce a color frequency heat map without migrant populations filtered out
PALM("DRB1*26F~28E~30Y", color=TRUE, filterMigrant=FALSE)

#example to produce a greyscale heat map with migrant populations filtered out
PALM("DRB1*26F~28E~30Y", color=FALSE, filterMigrant=TRUE)

}
\references{
Solberg et.al "Balancing selection and heterogeneity across the classical human leukocyte antigen loci: A meta-analytic review of 497 population studies". Human Immunology (2008) 69, 443–464
}
