% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PALM.R
\name{PALM}
\alias{PALM}
\title{Population Allele Locating Mapmaker}
\usage{
PALM(gdataset, motif, color = TRUE, filter_migrant = TRUE)
}
\arguments{
\item{gdataset}{Solberg dataset. Can be found at http://pypop.org/popdata/ under results.zip.}

\item{motif}{An amino acid motif in the following format: Locus*AA_1*AA_2*AA_3.}

\item{color}{Should the heatmap produced be in color? The default for color is set to true for a color heatmap -- for a map in grayscale, set color==FALSE.}

\item{filter_migrant}{Should OTH and migrant populations be filtered out of the dataset? The default for filter_migrant is set to true -- set this parameter equal to FALSE to keep migrant and OTH populations in the dataset.}
}
\value{
A .jpg file of the produced heatmap is output to the user's working directory.
}
\description{
Produces a heatmap from the manipulated Solberg dataset by using the Generic Mapping Tools (GMT) R Package, which is an interface between R and the GMT Map-Making software. GMT commands are called via a Bash script through using the gmt.system() function in the GMT R package.
}
